# Optimize the recent transactions query