@enobayram 2023-01-04
I've tested this PR by rerunning [the `/txs/events` queries from #109](https://github.com/kadena-io/chainweb-data/pull/109#issuecomment-1361666853):

```
$ time curl -H "Chainweb-Execution-Strategy:Unbounded" -v \
     "$CWDURL/txs/events?modulename=kaddex.aggregator&limit=10" \
   | jq '.[] | { height, name}'
< Chainweb-Next: KCgzMzQ0NjEzLCJOeDFaem0wdWZKa1hTZVVhR19BQ0xEc3VMckYwR25ua01MLXhhb2VLUDdRIiwyKSwwKQ

real	0m0,216s
```

``` json
{"height": 3346911, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3346268, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3346091, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3345816, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3345681, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3345525, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3345149, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3345044, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3344613, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
{"height": 3344613, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
```

```
$ time curl -H "Chainweb-Execution-Strategy:Bounded" -v \
    "$CWDURL/txs/events?modulename=kaddex.aggregator&limit=10&offset=2" \
  | jq '.[] | { height, name}'

< Chainweb-Next: KCgzMzQ2NTk5LCJrQjRlT25jMGIxVjl6TUYzTm5qd0U1ZVUyNHczWDJvWTBVbTJ5R3dXNFRjIiwwKSwxKQ

real	0m0,120s

<empty result>
```

```
$ time curl -H "Chainweb-Execution-Strategy:Bounded" -v \
    "$CWDURL/txs/events?modulename=kaddex.aggregator&limit=10&next=KCgzMzQ2NTk5LCJrQjRlT25jMGIxVjl6TUYzTm5qd0U1ZVUyNHczWDJvWTBVbTJ5R3dXNFRjIiwwKSwxKQ" \
  | jq '.[] | { height, name}'

< Chainweb-Next: KCgzMzQ1OTk1LCJjYiIsMCksMCk

real	0m0,131s
```
``` json
{"height": 3346091, "name": "kaddex.aggregator.PRIVILEGE-GUARD"}
```